environments/jdpg/requirements.txt
beautifulsoup4>=4.10.0
haystack-ai>=2.6.0
jupyter>=1.1.0
pdfplumber>=0.11.4
streamlit>= 1.39.0
watchdog>=5.0.0
azure-identity>=1.19.0